+++ .pre-commit-config.yaml
@@ -9,14 +9,14 @@ repos:
         name: isort (python)
   # black re-formats code to an opinionated standard, see https://black.readthedocs.io/en/stable/index.html for more
   - repo: https://github.com/psf/black
-    rev: 22.3.0
+    rev: 23.11.0
     hooks:
       - id: black
-        language_version: python3.9
+        language_version: python3.11
   # flake8 checks code to make sure it conforms to pep8 (we specify things to ignore in tox.ini)
   # see https://flake8.pycqa.org/en/latest/ for more
   - repo: https://github.com/pycqa/flake8
-    rev: 3.7.9
+    rev: 6.1.0
     hooks:
       - id: flake8
   # poetry hooks
