+++ article_rec_training_job/_0_fetch_events/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["GA4EventFetcher"]
+
+from .ga4 import GA4EventFetcher
